hw7.3

+++ src/main/java/ru/netology/JavaHomeWork7/service/StatsService.java
@@ -2,15 +2,15 @@ package ru.netology.JavaHomeWork7.service;
 
 public class StatsService {
     public long sumSales(long[] sales) {
-        int sum = 0;
+        long sum = 0;
         for (int i = 0; i < sales.length; i++) {
-            sum = (int) (sum + sales[i]);
+            sum = (sum + sales[i]);
         }
         return sum;
     }
 
-    public int averageNumber(long[] sales) {
-        int average = 0;
+    public long averageNumber(long[] sales) {
+        long average = 0;
         for (int i = 0; i < sales.length; i++) {
             average = (int) (average + sales[i]);
         }
@@ -41,7 +41,7 @@ public class StatsService {
 
     public int belowAverage(long[] sales) {
         int belowMonth = 0;
-        int averageSum = averageNumber(sales);
+        long averageSum = averageNumber(sales);
         for (int i = 0; i < sales.length; i++) {
             if (sales[i] < averageSum)
                 belowMonth += 1;
@@ -50,7 +50,7 @@ public class StatsService {
     }
     public int aboveAverage(long[] sales) {
         int averageMonth = 0;
-        int averageSum = averageNumber(sales);
+        long averageSum = averageNumber(sales);
         for (int i = 0; i < sales.length; i++) {
             if (sales[i] > averageSum)
                 averageMonth += 1;
